update go bindings gen script to have specific eth version

+++ scripts/generate_go_abi.sh
@@ -1,22 +1,17 @@
 #!/bin/bash
 #
-# Should be run from the project root and requires abigen from
-# https://github.com/ethereum/go-ethereum/
+# Generate Go ABI bindings for the contracts.
 #
-# Note that if you are switching between branches with different yarn.lock file you
-# may need to clean node_modules and yarn install --freeze-lockfile before running
-# this script.
+# Requires abigen from https://github.com/ethereum/go-ethereum/
 
 set -e
 
-SECONDS=0
-start=$(date +%s.%N)
+### Variables ###
 
-# Text colors
-WHITE='\033[1;37m'
-BLACK='\033[0;30m'
-NC='\033[0m' # No Color
+GO_ETH_VERSION="v1.10.26"
+GO_MODULE_NAME="github.com/chain4travel/camino-messenger-contracts/go/contracts"
 
+GEN_PATH="go/contracts"
 ARTIFACTS_PATH="artifacts"
 ARTIFACTS=(
     "contracts/account/CMAccount.sol/CMAccount.json"
@@ -29,19 +24,62 @@ ARTIFACTS=(
     "@openzeppelin/contracts/proxy/ERC1967/ERC1967Proxy.sol/ERC1967Proxy.json"
 )
 
-GEN_PATH="go/contracts"
+### END Variables ###
 
-echo -e -n "Cleaning node modules..."
-rm -rf node_modules && echo -e "${WHITE}done!${NC}"
+# Parse command-line arguments
+while [[ $# -gt 0 ]]; do
+    key="$1"
+    case $key in
+    --skip-yarn)
+        SKIP_YARN=true
+        shift
+        ;;
+    --help)
+        echo "Usage: $0 [--skip-yarn]"
+        exit 0
+        ;;
+    *)
+        echo "Unknown option: $1"
+        exit 1
+        ;;
+    esac
+done
 
-echo -e "Install dependencies..."
-yarn install --frozen-lockfile
+echo "Starting Go ABI generator..."
+SECONDS=0
+start=$(date +%s.%N)
 
-echo -e -n "Cleaning artifacts directory..."
-yarn hardhat clean >/dev/null && echo -e "${WHITE}done!${NC}"
+# Set workdir
+if [ -z "${WORKDIR}" ]; then
+    # camino-messenger-bot root folder
+    WORKDIR=$(
+        cd "$(dirname "${BASH_SOURCE[0]}")"
+        cd .. && pwd
+    )
+fi
+echo "cd $WORKDIR"
+cd "$WORKDIR"
 
-echo -e -n "Compiling contracts..."
-yarn hardhat compile --force >/dev/null && echo -e "${WHITE}done!${NC}"
+# Text colors
+WHITE='\033[1;37m'
+BLACK='\033[0;30m'
+NC='\033[0m' # No Color
+
+if [ -z "$SKIP_YARN" ]; then
+    echo -e -n "Cleaning node modules..."
+    rm -rf node_modules && echo -e "${WHITE}done!${NC}"
+
+    echo -e "Install dependencies..."
+    yarn install --frozen-lockfile
+
+    echo -e -n "Cleaning artifacts directory..."
+    yarn hardhat clean >/dev/null && echo -e "${WHITE}done!${NC}"
+
+    echo -e -n "Compiling contracts..."
+    yarn hardhat compile --force >/dev/null && echo -e "${WHITE}done!${NC}"
+else
+    echo -e "Skipping yarn install and hardhat compile..."
+fi
 
 echo "Generating Go ABI bindings..."
 for CONTRACT in "${ARTIFACTS[@]}"; do
@@ -73,6 +111,25 @@ for CONTRACT in "${ARTIFACTS[@]}"; do
     rm "$TMP_BIN_FILE"
     rm "$TMP_ABI_FILE"
 done
+
+# Go to the Go module directory
+cd "$GEN_PATH"
+
+# Clean go.mod and go.sum
+echo -e "Cleaning go.mod and go.sum..."
+rm -rfv go.mod go.sum
+
+# Run go mod init
+echo -e "Running ${WHITE}go mod init ${GO_MODULE_NAME} ${NC}..."
+go mod init ${GO_MODULE_NAME}
+
+# Get ethereum/go-ethereum
+echo -e "Running ${WHITE}go get github.com/ethereum/go-ethereum${GO_ETH_VERSION} ${NC}..."
+go get github.com/ethereum/go-ethereum@${GO_ETH_VERSION}
+
+# Run go mod tidy
+echo -e "Running ${WHITE}go mod tidy ${NC}..."
+go mod tidy -v
 
 end=$(date +%s.%N)
 runtime=$(echo "$end - $start" | bc -l | awk '{printf "%.2f", $1}')
